# fix(proxy): only fetch commit for root tree