This fixes a long standing performance issue with the file browser. There's still some room for improvement (the file list still doesn't load *instantly*), but I think that will require more involved work, such as a custom endpoint. For now, this makes it useable at least.

Fixes #350 

The issue was two-fold: the `tree::last_commit` function from `surf` currently has to go through the entire file (or tree) history to return the last commit. And we were calling this function for every folder in the tree.

The simple fix was to only call that function for the root folder, since we don't actually use the information for the sub-folders currently. We should also eventually make the `last_commit` function faster.

## Before
```
 INFO  proxy::http  > "GET /v1/source/revisions/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 17.081858ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 901.42691ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 886.55917ms
 INFO  proxy::http  > "GET /v1/source/blob/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 75.401932ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 123.660903ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 211.83489ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 166.012963ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 119.88884ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 1.293898509s
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 166.920815ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 883.429468ms
```

## After
```
 INFO  proxy::http  > "GET /v1/source/revisions/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 21.978566ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 22.852715ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 19.194392ms
 INFO  proxy::http  > "GET /v1/source/blob/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 75.114188ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 19.321996ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 19.673202ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 18.543029ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 18.588068ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 18.85306ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 20.359387ms
 INFO  proxy::http  > "GET /v1/source/tree/464b300a3677d0ca5410cdb21f42c3ca54f3939a.git HTTP/1.1" 200 19.056329ms
 ```